+++ bake.rb
@@ -11,6 +11,6 @@ def after_gem_release_version_increment(version)
 	context["utopia:project:update"].call
 end
 
-def after_gem_release(tag_name:, **options)
-	context["releases:github:release"].call(tag_name)
+def after_gem_release(tag:, **options)
+	context["releases:github:release"].call(tag)
 end